added langsmith, unittests and defaults for local groq llm calls
