+++ agentic-trader/decision_chain.py
@@ -2,7 +2,7 @@ from langchain.prompts import ChatPromptTemplate
 from llm_router import get_llm
 
 
-def run_agent(ohlc_data, volume, provider="openai"):
+def run_agent(ohlc_data, volume, provider="groq"):
     llm = get_llm(provider)
 
     template = ChatPromptTemplate.from_messages([
